Database Backup
Added the script I use to perform a backup of my databases.

I have a core for Vanilla, one for The Burning Crusade, one for Wrath of the Lich king and one for Cataclysm so the script needs to export them all without the logs (Vanilla) and hotfixes (Cataclysm) databases.

The script will export and compress all databases that do not contain the word 'world', 'logs' or 'hotfixes' anywhere in their name.

My config will save 168 archives and when it reaches that point will delete the oldest file on next backup.