Rewrite
I finally got around to rewriting the bash script for AzerothCore. In it's current state, the script does nothing. It will take a while until it's back to a usable state, and even longer to finish it.

The idea is to clean it up, iron out some issues and improve upon it.

Being able to store the version of the installed client data files and update if required is one minor thing I will be adding. It will also give the user an easy way to tell the script what the required version is. There are many uses for being able to update the XML-file from within the script but this is the big one.