AzerothCore: Menu
Began work on a system based on a visual menu, using the package whiptail, that will be used when no parameters are passed to the script. It may help someone who needs it, however I will personally not use the menu but still want it to work just like when not using it.