AzerothCore: Update
The script will now clone, or fetch, the latest version of the repository, including enabled modules, and compile it.

It will work just as intended by using the simple script parameters or by using the visual menu.

I decided not to use a progress bar when performing tasks with the menu but instead just to use the terminal. It was easier and couldn't find a way send the proper percentage to a progress bar.