AzerothCore: Client Data
Added the ability to download, and update, the client data when required. Should work just fine no matter if done through the menu or terminal. The trigger for downloading the files depend on the 'required_client_data' and 'installed_client_data' being different. When it's done downloading and unpacking the files, it will update the 'installed_client_data' value with that of 'required_client_data'.

It also allows a user to update the data version without me having to do it first. Simply changing the 'required_client_data' whenever a new one is released is sufficient.