AzerothCore: Updates
Added the section for generating the start and stop scripts, along with the auto-restarter for each process. It will only generate those that the user needs just like the previous script.

Added a check to the client data files. If, for some reason, the set version in the xml file is the same as the required one but at least one of the folders is missing - it will download them all again.