AzerothCore: Fixes
Fixed an issue when enabling and disabling the Eluna module. In certain cases it would disable the module when not actually changing the value itself within the menu. This should, hopefully, no longer happen.

Reduced the size of all available elements. It might not look as pleasing but I want to minimize the required space the menu takes up.