AzerothCore: Database
Started work on the database part of the script. I need to find a pretty way of doing it no matter if it's initiated from the menu or the terminal.

The systems differ a little, the menu allows the user to import each database separately while the terminal does it based on if the auth and/or worldserver is selected.

I'm also doing some cleaning up and minor tweaks to hopefully manage things smoother.