AzerothCore: Database, Configuration and fixes
Added the ability to import and update the databases via the menu. Added the forgotten call to the function of updating the databases.
Added the ability to update all configuration files via both the menu and the terminal.

This is really a first draft and will be improved over time. I still need to add the function to handle start and stop of the server but that is the easy part.