AzerothCore: Major changes
Unfortunately I decided to remove the menu for the time being. I want to do a full rewrite to the point where I'm happy with the results.

Currently the script works just like it should from the terminal. I am fairly pleased with how it turned out, but I will be adding the menu again.