Powershell: MySQL Backup
This is the script I will be using to perform backups of my databases. Thea idea is that I will use a server running Windows 2019 and uploading all exported data to my second google drive. I could do this using Bash and Ubuntu, with tougher work,, but it gave me an excuse to work with Powershell.