Powershell and Bash: Database Backup
Removed the powershell script. It was a mistake to even make and I will not be using it.

Added a rewrite of the database backup script in bash. It has a variable called backup_type that can either be local or gdrive. If it's set to local, it will simply create archives in a folder called database in the same folder as the script. If it's set to gdrive, it will create archives in either /root/gdrive/database or if not used as root in /home/username/gdrive/database. This is a script I'm going to be running every hour and thanks to Google I don't have to worry about the possibility of a failing drive in my setup.