AzerothCore: Fixes
Fixed the issue where the script doesn't update the databases. It was an oversight on my part and I honestly don't know how I missed it.
Fixed, or added, the part where the script imports any custom content to the world database.
Started working on the menu now that the rest is, as far as I know, nearly finished.