AzerothCore: Revert
I feel like an idiot but I want to try to do this without the use of whiptail. It should be easier and nicer to manage. I found out that I'm not very pleased with the looks or the flashes that occur when you change between menus quickly. Epileptic anyone?