AzerothCore: Changes
Changed all instances of echo to printf. There's no real reason for this except just that I felt like it.
Added all available colors as variables. The colors are, technically, optional and the script will not stop if that include is missing.

These changes are in relation to the menu I'm working on, making it easier for me in the end.