AzerothCore: Menu additions
Added the ability to view and change the settings related to the databases via the menu.
I opted into having the password viewed in the menu be replaced with asterisks, but changing the password is not hidden. I could do it but there's not really a point since if someone has access to the script, they have access to the xml file anyway.