AzerothCore: Menu - Config options
Added every current config option that the user can change. It will only list them right now, the next step is implementing the system that allows the user to modify the values.

I'm not 100% on any of the grammar. I wish I could find the perfect way of phrasing everything, but maybe I'll fix that piece by piece over a longer time.