AzerothCore: Menu
Added the ability to modify the first 8 options of the menu. Unfortunately, due to limitations of the terminal, it's not possible to edit the 'Message of the Day' this way. The only way to edit that is by modifying the xml-file manually. It will prompt with this information if selecting it in the menu too.
Added more selectable game types and realm zones in the menu.