AzerothCore: Updates
Added the list of possible realm zones to the configuration. I will implement a system in the Bash script that will only accept specific values in certain places like game type or realm zone. Should a value be invalid, it will reset to their default value.
Updated the Powershell script more, still far from usable but it's being worked on.