AzerothCore: Fixes and additions
Fixed an issue with the Bash script that was caused by my change of the variable name.
Added a huge chunk of the configuration to the Powershell script. It will, just like the Bash script, reset values that are invalid. I am basing the Powershell script on the Bash script, trying to keep them as close to each other as possible.