AzerothCore: Restructure
Rewrote a part of the configuration. Changed many values to be false or true instead of 0 or 1 to be easier understood by the.. lets just say average user.
Merged the two parts of the database script together since they all use the same variables and are usable in the exact same way.
Removed the menu temporarily while I work it through, the idea I had in mind takes more effort than I thought and it is just easier to start from scratch and do it right all the way through. It doesn't take that long to write.