AC: Packages and menu
Added a function to install Git when cloning the source code. My reason for this is if a user only wants to clone the source, for whatever reason, but not compile it - they shouldn't need to install all those packages that are required to compile the source.
Added the part of the menu that manages the modules, downloads the repository/repositories and that compiles the code.