AC: Menu
Added the section that handles the databases.
Added the section that handles the starting and stopping of the auth and world server. I'm unsure on how I want the starting to look since it's instant and doesn't give a hint about how it's being started but for now a simple sleep will do.