AC: Configuration
Added a check to make sure installed_client_data isn't a higher value than required_client_data.
Added a way of handling the client data automatically. Setting required_client_data to 0 will cause it to attempt to grab the latest possible release automatically and determine if it needs updating. This is not set to 0 by default as it may, or may not, cause issues. If issues occur with this solution, setting it to a value yourself is the best option.