AC: Configuration
Added the ability to update the required client data files version by command and by using the menu. The command ./azerothcore.sh client will update the version if there is one. The menu will do the same.
For the time being I will not be adding it to the 'setup' and 'all' commands until I've tested what happens if it can't fetch the version, for example if it can't connect to the repository.