AC: Configuration (Menu)
Added the ability to change every available option. I need to go over the integer options and figure out a way of handling them. The ideal solution is to print the current value, which for instance the database options does, but without it accepting values outside a certain range. If I add this print to, for instance, the max level value it will accept values > 80 for some reason.