AC: Menu
Rewrote a huge chunk of the menu. It will no longer auto accept a single digit in the menus but the user has to input a number followed by the return key. If no value is inputted prior to pressing the enter key, it will go back to the previous menu. It also allows me to reduce the amount of 'pages' used by the configuration options and therefore also clean it up quite a bit. I also decided to add spaces in various menus to make it look much neater.
I did decide to keep the 'input a number in the options to switch between various options for that option' since it also saves me a lot of pages.