AC: Modules
Added the ability to copy lua scripts to bin/lua_scripts. Any files placed inside the folder called 'lua' in the same location as the script will be copied over. For now it will delete the lua_scripts folder and recreate it. I decided on this to support removing lua scripts.