Fixes, changes and additions
Added support for my module called Assistant. Not all settings are in yet, but most of them are. I will be adding the rest of them in the near future.
Removed the menu. It is just so much effort for such a small reward. Especially since I don't even use the menu myself.
Moved the invalid argument notification to it's own function so any changes to it in the future are easy.