Removed
Removed the powershell script for AzerothCore. I'm not going to finish it since I'll never be using it and I simply don't have the time or energy to spend on it.
Removed the LUA scripts since I can't even use the Eluna module, it won't compile, and I did make my own module to replace it.

The module is located at https://github.com/tkn963/mod-assistant