LUA: AzerothCore
Added the weekend double experience and reputation.
Added the spells on level up and on login.
Added spawning new characters in Stormwind and Orgrimmar.

I am keeping them all separate so it's easy to just remove a script if it's no longer needed. There won't be any unnecessary files because of this.