Quest loot
Updated the chance of items dropped from creatures that are normally looted from objects on the ground to be a viable way of obtaining the item. I also added them to the tooltip of the creature.

This is all to make it more friendly towards multiboxers.

The items were already there, with a chance of 1%.