Overrides
Added a way to override server scripts whenever the source code is updated.
This is obviously not ideal, but it does get the job done.

I wanted to do this because there are some quests that only trigger for the player doing it rather than the group they're in, I wanted to try to change those scripts without modifying the entire C++ file.

I will be adding more quests to the overrides, of course.