Interfaced
Removed the BartenderToggle addon.
The Interfaced addon now handles what the BartenderToggle addon did.
Added key bindings to move the UI elements into position and to toggle the bartender visibility.

I can't find a way to automatically moving the UI elements into position when they are moved, for example when entering or exiting a vehicle, so the key bind is there to handle it until I can find a way.