Interfaced
Removed the key binding for moving UI elements.
Added an OnUpdate script that will move the UI elements into position roughly every second IF they are not in the correct position.

I need to rework the handling of the party member frames. We don't want the addon hammering updates into frames that do not need to be updated.