Rewrite
Began a complete rewrite of the bash script for azerothcore.
It's far from usable, but with the automatic updater will help me a lot.

As written in the readme, do not download the files inside the core folder. The script, azerothcore.sh, will download those automatically to trigger updates.

I use many different systems to run my azerothcore, this means I don't have to upload the changed scripts to all those systems.