Source code
Added the functions used to manage the source code. It will download, or update, the repositories enabled in the options file, compile it and then download the client data files. It will also check for a newer version of the client data files before downloading them. This means you don't have to do anything when there are updates to those files. It will also create bash files that are used to start and stop the server, including automatically restarting on crashes.
Removed the case at the end, went with a simple if instead since I'm handling this a lot differently than before.