Modules
Added support for all modules. Modules that have their own tables will be handled differently and those tables will be removed when disabling the module. The only exception is the recruit a friend module.

Added support for importing custom content inside the sql/world folder that should be created in the same folder as the script is in.