Crontab
I recently changed my system to have 1 VM running auth, 1 VM running regular azerothcore, 1 VM running my wrath of the vanilla variant and 1 last VM running my wrath of the vanilla variant as a PTR that I use when I'm working on the code. These are the crontabs I use for that.