AzerothCore
Removed the automatic updater for the azerothcore bash script. I don't update it often enough to suffer from having to upload changes to each machine I use it on. I ran into several issues when using the updater that I can't figure out and simply don't have the time or energy to give a lot of focus.