Crontab
Added the commands I use that executes when the system is restarted. Occasionally I do reboot my 2U servers and most of my containers are set to start on boot - this helps me by not forcing me to log onto each system and start the processes. It will merely sleep for 300 seconds (5 minutes) to make sure the database servers are fully up and running and then start my auth server and all my world servers.