Rewrite
I did a full rewrite of the backup script. There are still certain issues that could happen if not properly handled.

Using the google drive storage option assumes that it has been set up for the current user at $HOME/gdrive (~/gdrive). If using this option and not having it set up properly, I will not be offering support.

I am still in the process of testing this, but hopefully there are no major issues.