VuhDo
Removed all errors printed by the buffwatch. It will no longer print a message if used in combat or if there are no buffs to cast.
Changed the handler for HoTs to have one icon just above the player name, which I use to show debuffs only.
Added a space between the health and mana of a players frame to give it a little more retail feel.