Features
Added the ability to disable leaving your group when you log off. It only affects regular groups, not raids or dungeon finder groups. I was getting tired of leaving the group whenever I relog one of my characters. It should be a config options in AC, perhaps I'll make a feature request in the future.
Allow setting rested experience rate to 0. I don't like rested experience and want it to be disabled on my end.