Minimal
This is my work towards having a more minimalistic version of the script that doesn't handle modules or anything else flashy.

It's also a place where I can begin to implement a menu that'll help users navigate various options.

The regular script will get the same menu, but modified to support modules, when this one is done. It's not anywhere near usable at this point, though, so don't try to use the menu yet.