Fixed
Fixed an issue using the boost package set for Ubuntu. Using Ubuntu 21.10, the same package is used as with Debian 11. I have yet to test it under Ubuntu 20.X to make sure it's changed in that version too but this solves the issue on my end for the time being.