Default values
Updated the default value of the required client data files version. The script does grab the latest available version on it's own, regardless.
Changed the default value of the ahbot modules account id and character guid to 1. It should never be 0 if the module is used, makes no sense to have that as the default value.