Preparation
Added a few lines of code that will be implemented eventually. With the changes to the core that allows you to compile only one part of the core, like not wasting time compiling both auth and worldserver on a system that only runs the authserver, this will work with that. Unfortunately there are errors when you try to compile using these new settings so I'm leaving it commented out for the time being.