Assistant
Added a variant of my Assistant module. It's not complete yet, I need to add the profession feature, but it's usable. The downside is that you have to start the server and then restart it when it's up to reload the changes to the item_template table. You only have to do so the first time unless you copy the SQL queries I've added near the top and run them on your world database before starting the server.