Ubuntu 20.x
Removed support for Ubuntu 20.04 and 20.10. It doesn't work as it is but require some custom changes to work on those versions. I haven't tested 21.04 or 21.10 since all I use is Debian but I might boot up a VM to see if it still works. I created this script for myself so I don't care if I have to revoke support for all distributions other than Debian.