AoWoW
Dramatically reduced the time it takes to update the aowow data by removing entries that update the same tables as another entry.
Added a user input feature for Grandold. If the script is run like './sync.sh cron' it will continue to sync without waiting for the input.
Added a print that will tell the user how long the update took to finish.