Fixes
Updated the sync script to include some color and to tell the user what step it's starting so they know where in the chain they are.
Updated the azerothcore script to include the new config option I added to the account bound module.
Removed the extra spawn points I added for the assistant.
Added arrows and bullets to the free reagents script.
Added spawn points for the heirloom hoarder for the account bound module to each starting zone.