Server shutdown
The script will now tell the world server to shutdown instead of save and killing it. It should hopefully help in preventing a rollback. I've tested it on my end and it works just as it should but I can't guarantee that everyone will end up with the same results.