AzerothCore
Replace the progressive module with the progression module. It's not yet public but meh, it helps me until then.
Added the ability to copy custom dbc files from a folder called dbc in the same place as this script.
Updated the ability to import custom SQL files to a folder called sql in the same location as this script instead of sql/world.