Overhaul
Overhauled the script for AzerothCore because it's far too complicated for no real reason.
The only downside is that now all options are stored in the file itself. Any option not set in a variable is set to what I want it to be.
I also decided to remove support for all distributions except Ubuntu 22.04. I will be adding Debian 12 eventually.
This reduced the number of lines from 3489 to 1045.