Fixes
Updated the config for the ah bot to actually use sell price when buying items rather than buy price. This config option is for some reason reversed within the module and you have to enable the option to actually disable it.
Added a log to the world server. This helps in finding out the reason if and why the world server crashed.