AoWoW
Rewrote the script for syncing AoWoW to allow full automation. It will drop all tables every time but that's the idea, to make sure there are no issues. I don't sync it often so I prefer to leave it this way.
The files inside the sql folder are required to prevent errors.