Playerbots
Add a version of the script that includes the playerbots module. Seeing as it requires a special fork to function, that is up to the user to change (in the config.sh file) by pointing it to the fork instead of the official repo. I decided to leave it as the default value since I point it to my own fork with the playerbots changes merged into it.