Cluster
Added support for using a cluster. It's still being worked on and will probably need tweaking along with a few other features but it does work and can be properly tested.

I know there's not a lot of documentation on using any of this but that's not very important to me at this stage. I'll probably add it in the future.