DataDir
Added support for defining the path to client data files. This is helpful if multiple realms, or nodes in a cluster, is running on the same system so it doesn't take up extra space storing the data files for each one.
Removed the defined data version in the database along with this since it has to be stored with the data files in order to be read by multiple instances of this scripts.