Modules & Cluster
Remove the piece of code that deletes the build folder when a module is removed. This is completely unnecessary and only prolongs the process. Should issues arise by the removal hopefully the make clean will pick up the pieces.
Updated the health check port for each cluster node in an attempt to make sure there are no conflicts.