README
This is the first bit of the documentation. I'm horrible at instructions but hopefully it's clear enough to be worthwhile.
I'm not done with the config section yet, not even half the options are documented, but I need to deal with some other stuff so it'll have to do for now.