Config
Update config options based on set patch. Modules that only include features available in later patches of the game shouldn't be installed if the patch doesn't allow it
The ahbot module shouldn't post items with a higher item level than available in the patch. I need to go through and tune it for patches before 3.0