Fixes
Fixed an issue where it updated worldserver conf with true/false instead of 0/1. It works either way but consistency.
Enable the avoid aoe feature of playerbots by default. It's still experimental but worth trying.
Removed the package golang-go when using cluster. Ubuntu 22.04 includes a version that's too old so manual installation is required.