Cleanup
Removed all settings that I either never change or that is always the same across all my realms. This script is here for me and I prefer not to add a lot of stuff that I'll never even use. I do change some settings based on the progression patch but it's better and easier just to let the script handle that for me.