Additions
Added support for Eluna.
Changed git pull to git fetch.
Added a message showing how long certain processes took to complete. This includes downloading and updating the source code, compiling the source, and importing the database files.
Added a flag in cmake when using playerbots to stop all those warnings that dramatically increase the compile time when updating the core.