World Reset
Added the ability to drop all tables from the world database, which is useful when you need to reset it. It will not touch the auth or characters databases, and will not do anything if build.world is disabled.

I will probably be adding additional parameters to this so it's possible to reset the source code as well as drop the tables. I haven't decided if the entire source folder should be deleted or if it should just run make clean though.