Playerbots
Reworked the way I handle playerbots. The number defined in the database is set as MaxRandomBots and MinPlayerBots is set to half of that value. The number of accounts is set to the value from the database divided by 9 and 1 is added to the result. It's hard to explain but it does what I want without me having to change several settings when increasing or decreasing the number of bots.
Removed the talent changes for the bots for now. For some strange reason some spells weren't picked up by heal prediction with the changed talent trees and I don't want to play without it.