Complete rewrite of the script
There are still parts not yet done. Start/Stop/Restart/Reset still needs to be implemented but everything else should work as expected.

This code is cleaner, better, has improved readability and makes making changes a whole lot easier.

It'll also print how long each step took. It's a feature I've missed for a long time and it's good that I finally took the time to do all this.